requirements_webapp.txt
# ADF Monitor Pro - Enterprise Web Application
# Core Dependencies
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0

# Database and Data Processing
sqlite3  # Built into Python
sqlalchemy>=2.0.0
pymongo>=4.5.0  # For MongoDB support
psycopg2-binary>=2.9.0  # For PostgreSQL support

# Azure Integration
azure-identity>=1.14.0
azure-mgmt-datafactory>=3.0.0
azure-storage-blob>=12.17.0
azure-keyvault-secrets>=4.7.0
azure-monitor-query>=1.2.0

# AI/ML Providers
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0
transformers>=4.35.0
torch>=2.0.0  # For local models

# Web Framework and UI
streamlit-aggrid>=0.3.4
streamlit-option-menu>=0.3.6
streamlit-plotly-events>=0.0.6
streamlit-ace>=0.1.1
streamlit-authenticator>=0.2.3

# Notifications and Communication
requests>=2.31.0
python-dotenv>=1.0.0
smtplib-ssl>=1.0.0
slack-sdk>=3.22.0
teams-webhook>=0.1.0

# Security and Authentication
cryptography>=41.0.0
jwt>=1.3.1
bcrypt>=4.0.1
python-jose>=3.3.0

# Monitoring and Logging
prometheus-client>=0.17.0
loguru>=0.7.0
structlog>=23.1.0

# Testing and Development
pytest>=7.4.0
pytest-mock>=3.11.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Background Tasks and Scheduling
celery>=5.3.0
redis>=4.6.0
apscheduler>=3.10.0

# Data Validation and Serialization
pydantic>=2.4.0
marshmallow>=3.20.0
jsonschema>=4.19.0

# Performance and Caching
cachetools>=5.3.0
memory-profiler>=0.61.0
line-profiler>=4.1.0

# Configuration Management
pyyaml>=6.0.1
toml>=0.10.2
configparser>=5.3.0

# Development Tools
jupyter>=1.0.0
ipython>=8.15.0
autopep8>=2.0.0
